backend/requirements.txt
Django>=6.0
djangorestframework>=3.16.0
django-cors-headers>=4.9.0
Faker>=24.0.0
pandas>=2.0.0
openpyxl>=3.1.0

